Arduino (Motorlib_v7) - Speed up sensor read time

+++ Arduino/motorLibv2/Arduino_SharpIR-master/src/SharpIR.cpp
@@ -1,8 +1,8 @@
 #include "SharpIR.h"
 
-uint8_t SharpIR::getDistance( bool avoidBurstRead )
+unsigned double SharpIR::getDistance( bool avoidBurstRead )
 	{
-		uint8_t distance ;
+		unsigned double distance ;
 
 		if( !avoidBurstRead ) while( millis() <= lastTime + 20 ) {} //wait for sensor's sampling time
 
@@ -12,7 +12,7 @@ uint8_t SharpIR::getDistance( bool avoidBurstRead )
 		{
 			case GP2Y0A41SK0F :
 
-				distance = 2076/(analogRead(pin)-11);
+				distance = 2076.0/(analogRead(pin)-11.0);
 
 				if(distance > 30) return 31;
 				else if(distance < 4) return 3;
@@ -22,7 +22,7 @@ uint8_t SharpIR::getDistance( bool avoidBurstRead )
 
 			case GP2Y0A21YK0F :
 
-				distance = 4800/(analogRead(pin)-20);
+				distance = 4800.0/(analogRead(pin)-20.0);
 
 				if(distance > 80) return 81;
 				else if(distance < 10) return 9;
@@ -32,7 +32,7 @@ uint8_t SharpIR::getDistance( bool avoidBurstRead )
 
 			case GP2Y0A02YK0F :
 
-				distance = 9462/(analogRead(pin)-16.92);
+				distance = 9462.0/(analogRead(pin)-16.92);
 
 				if(distance > 150) return 151;
 				else if(distance < 20) return 19;
